requirements.txt
loguru
PyYAML

aiogram